feat: RevealSectionMotion으로 모션 추상화 및 모션 적용
