+++ app/(sub)/capsule-detail/_components/caption-section/caption-section.css.ts
@@ -1,5 +1,5 @@
 import { themeVars } from "@/shared/styles/base/theme.css";
-import { screen } from "@/shared/styles/tokens/screen";
+
 import { style } from "@vanilla-extract/css";
 
 export const container = style({
@@ -7,9 +7,6 @@ export const container = style({
   display: "flex",
   flexDirection: "column",
   gap: "1.6rem",
-  ...screen.md({
-    margin: "1.6rem",
-  }),
   width: "100%",
   color: themeVars.color.white[85],
   background: themeVars.color.gradient.darkgray_op,
